[AbstractPls] abstraction attempt, needs testing
