+++ FolderBehaviour.java
@@ -1,0 +1,7 @@
+public interface FolderBehaviour {
+
+    public static final String DONE = "Finished";
+
+    public void doCreate(String filePath);
+    
+}
